Move Watchlist to security_admin and other clean-up Some references to the old admin accesses were still hanging around, so we've removed those and moved the watchlist to a more logical place.
