+++ uber/templates/registration/menu.html
@@ -3,7 +3,7 @@
     "form?id={id}", "Attendee Data", True,
     "shifts?id={id}", "Shifts", attendee.staffing,
     "history?id={id}", "History", True,
-    "watchlist?attendee_id={id}", "Watch List Entry", attendee.banned and c.HAS_DEVTOOLS_ACCESS,
+    "../security_admin/watchlist?attendee_id={id}", "Watch List Entry", attendee.banned and c.HAS_SECURITY_ADMIN_ACCESS,
     "index", "Return to Attendee List", c.PRE_CON,
     "index?search_text={id}", "Return to Attendee List", c.AT_THE_CON,
     "index", "Return to Attendee List", c.POST_CON
